Début panier

+++ application/models/Produit_model.php
@@ -4,28 +4,19 @@ defined('BASEPATH') OR exit('No direct script access allowed');
 
 class Produit_model extends CI_Model {
 
-    public function insert_list($data) {
-        // On spécifie les valeurs à inserer dans la table        
+    public function create_product($data) {
+        // Insert un produit in bdd      
         $this->db->insert('produits', $data);
     }
 
-    public function read_list() {
-        $this->db->select('*')
-                ->from('produits')
-                ->join('categories', 'categories.cat_id = produits.pro_cat_id');
-        $result = $this->db->get()->result();
+    public function read_products() {
+        // Lis toutes les lignes de la table produits
+        $result = $this->db->get('produits')->result();
         return $result;
     }
 
-    public function read_cat() {
-        // On sécifie qu'on veut toutes les données dans la table
-        $result = $this->db->get('categories')->result();
-        // ensuite, on envoi le résultat
-        return $result;
-    }
-
-    public function read_one_product($id) {
-        //$result = $this->db->get_where('produits', array('pro_id' => $id))->row();
+    public function read_by_product($id) {
+        // Lis les informations d'un produit
         $this->db->select('*')
                 ->from('produits')
                 ->join('categories', 'categories.cat_id = produits.pro_cat_id')
@@ -33,23 +24,28 @@ class Produit_model extends CI_Model {
         $result = $this->db->get()->row();
         return $result;
     }
-    
 
-    public function update_list($id, $data) {
-        // On doit spécifier en premier lieu l'identifiant de la ligne à modifier
+    public function read_categorie() {
+        // Lis toutes les lignes de la table categories
+        $result = $this->db->get('categories')->result();
+        return $result;
+    }
+
+    public function update_product($id, $data) {
+        // Modifie le produit sélectionnée
         $this->db->where('pro_id', $id);
-        // ensuite les valeur à changer dans la table
         $this->db->update('produits', $data);
     }
 
-    /**
-     * 
-     * @param type $id
-     */
-    public function delete_list($id) {
+    public function delete_product($id) {
+        // Supprime produit in bdd
         $this->db->where('pro_id', $id);
         $this->db->delete('produits');
-        unlink(FCPATH . 'assets/image/' . $id . '.' . 'jpg');
+        // Supprime image in dossier
+        $filename = FCPATH . 'assets/image/' . $id . '.' . 'jpg';
+        if (file_exists($filename)) {
+            unlink(FCPATH . 'assets/image/' . $id . '.' . 'jpg');
+        }
     }
 
 }
